pyproject.toml
[project]
name = "auth-service-poetry"
version = "0.1.0"
description = ""
authors = [
    {name = "Boris Averin",email = "89068157313@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.3,<3.0.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-extra-types (>=2.10.4,<3.0.0)",
    "ruff (>=0.12.2,<0.13.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
]

[tool.poetry]
packages = [{include = "auth_service_poetry", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
